requirements.txt
appdirs==1.4.3
#Brlapi==0.6.6
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
packaging==16.8
pyparsing==2.2.0
requests==2.21.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
flask-sqlalchemy==2.3.2
PyMySQL==0.9.3
